chore(boxes): adding an init command for an empty project (#7398)
Quick PR to add a `npx aztec-app init` command that provides a `nargo
new`-like feature. This will be used in tutorials and specially in the
EthCC workshop.

---------

+++ .devcontainer/scripts/onCreateCommand.sh
@@ -13,14 +13,17 @@ fi
 if ! grep -q "alias sandbox" ~/.bashrc; then
     echo "alias sandbox=\"npx aztec-app sandbox\"" >> ~/.bashrc
 fi
+corepack enable
 
-source ~/.bashrc
-yes | npx aztec-app -t $TYPE -n $NAME -s
-mv $NAME/* $NAME/.* .
-rm -rf $NAME
+if [ "$TYPE" != "sandbox_only" ]; then
+    source ~/.bashrc
+    yes | npx create-aztec-app -t $TYPE -n $NAME -s
+    mv $NAME/* $NAME/.* .
+    rm -rf $NAME
+    
+    yarn
 
-yarn
-
-npx -y playwright install --with-deps
-yarn add @aztec/builder
-yarn prep
+    npx -y playwright install --with-deps
+    yarn add @aztec/builder
+    yarn prep
+fi
